requirements.txt
inpainting==0.0.5
torch>=2.0.1
torchvision
numpy
tensorboard==2.5.0
Omegaconf
Pillow
scikit-learn
matplotlib==3.3.1
setproctitle==1.1.10
SimpleITK
albumentations==0.5.1
einops==0.3
pandas==1.2.2
pandasql==0.7.3
art==5.1
nibabel==3.2.1
byol_pytorch==0.5.6
visdom==0.1.8.9
tqdm
pynacl
prettytable
torchmetrics
blobfile
nibabel
